third_party/WebKit/LayoutTests/platform/android/fast/workers/worker-onerror-08-expected.txt

This tests that exceptions in setTimeout triggers 'worker.onerror'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Page-level worker.onerror handler triggered:
PASS errorEvent.message is "Uncaught Error: Exception in setTimeout callback"
PASS stripURL(errorEvent.filename) is "[blob: URL]"
PASS errorEvent.lineno is 3
PASS errorEvent.colno is 19
PASS errorEvent.error is null
PASS successfullyParsed is true

TEST COMPLETE
